Function changers::logic::change::make_change

pub fn make_change(
    n: f64, 
    coins: Vec<f64>
) -> Result<HashMap<String, i32>, LogicError>

Make change

Errors

Returns a LogicError::Unreachable when the desired amount cannot by reached using the given coins.

Examples

use changers::logic::make_change;

let n: f64 = 10.0;
let coins: Vec<f64> = vec![1.0, 2.0, 5.0];

let change = make_change(n, coins).unwrap();
assert_eq!(change["5"], 2);

Trying to make change for an unreachable amount

use changers::logic::make_change;

let n: f64 = 3.0;
let coins: Vec<f64> = vec![2.0, 4.0];

let change = make_change(n, coins);

assert!(change.is_err());
assert_eq!(
    change.unwrap_err().to_string(),
    "Amount cannot be reached with the given set of coins"
)