Function changers::logic::coins::normalize

pub fn normalize(
    n: f64, 
    coins: &Vec<f64>
) -> Result<(u32, Vec<u32>), LogicError>

Normalize a vector of coins

The vector of coins is normalized with the factor 10^n, where n is the amount if decimal places of the longest coin.

Errors

Returns a LogicError::Unreachable if the precision of the desired amount is greater than the precision of the longest coin.

let n: f64 = 3.5;
let coins: Vec<f64> = vec![1.0, 2.0, 5.0, 10.0];

assert!(normalize(n, &coins).is_err())